Crashing the game
I wrote this on my terminal, pasted this onto the game and applying it...
it crashes, both on Firefox and on Chrome.

I probably have a bug on my code that is causing the crash, but the game
should behave better anyway, so I'm commiting this to open a bug report.